Moved the ToolstripCommand to the Controls directory.
